Rudimentary health check and framework for a server
This is a proof of concept to get a health check up and running for use by other software that will monitor this application.

The server portion is only partially implemented (enough to log that it isn't starting).

Mostly this commit is intended to be a code review and explanation of the direction we could (should?) take with the codebase.