Flipping the library exports to export the class
Libraries were exporting a function that required use of a DI system in order to create the objects.  This commit pulls the container-related code out of the libraries, making them more usable to a wider audience.  All of the container-specific code is now in dependencies.js and whatever uses dependencies.js (like bin/server.js).