Use async processes, starting with certificates and key files (#36)
* Partially implement promises for Web Server

The server works but the tests need updating.

* Pulled in a synchronous promise testing library

This made most of the tests pass without any other problems.  Finsihed
up converting the fs.readFileSync() to fs.readFileAsync() and away we
go!

* Reworking promises to trigger only on .startServer()

Major changes to add contracts that will be added to the promise when
the server is created.  They are not added on the fly, but just enter a
queue and will show up magically when .startServer() is called.  This,
of course, broke tests.

* Adding tests for the library

* Removing unneeded file

* Properly excluding everything under spec/

* Unneeded