Implement significantly stricter linter (#71)
This adds a much more strict .eslintrc (also converts it to YAML).  I've
ignored node_modules/ and coverage/, but the rest of the files now get
scanned when running "npm test".

There were many little changes that needed to get made so the linter
would not find more problems.  It's a little strict, like jslint, but
that means our code will be more uniform.