Rename secureHash and split off encoding (#94)
* Rename secureHash and split off encoding

Renames secureHash to just hash in preparation for future work.  The
secureHash.hashAsync method is renamed to hash.deriveAsync because it is
a key derivation function.  In the future there will be a simpler
version that just does a single has and it will be hash.hashAsync.

Split off the encoding to make it more modular, key it off config, allow
it to use other modules and to allow other modules to use it directly.

* I had these backwards

Moreover, I was totally blind when I ran the tests on my side and missed
the several lines of errors output to my screen.  Apologies.