Allowing multi-part IDs for storage (#98)
* Allowing multi-part IDs for storage

You enable them by passing multiple ID hash config objects to the
storage service factory.  The storage service will then also receive an
array for the `id` of the record to manipulate.  Each part of the array
will map to the hash at the same position.  If there's more IDs than
hashes defined, the hash definitions will loop.

This also removes some unnecessary dependencies in a couple modules and
consolidates code for duplicating objects, which may help prevent
problems with objects changing in memory (though that's not supposed to
happen in the first place).

* Forgot unit test

* Fixing linting issues

I wonder why this wasn't checked when I saved it.  Either my check is
broken or else I somehow totally ignored it.

* Additional clone tests