Updating the validation middleware
This renames the middleware to indicate that it only validates the body.
There will be a new middleware that validates the query string, which is
why this change needed to be made.

I also extracted the mock middleware into its own mock file and had to
then update four route tests to inject the mock into the container.

The majority of the rest of the changes are simply updates to use the
new middleware's name.