Better error handling with error objects and formatters
The errorResponse module now can make ErrorResponse objects.  If you
pass an ErrorResponse object to `res.send()` it will be formatted as a
standard error output.  To do that, all of the formatters needed to be
rewritten because I could not inject middleware or a prefilter of any
sort into Restify.

This also breaks all of the other formatters out into separate files,
making testers much easier.

Restify also (weirdly) never sets the Content-Length header and relies
on the formatters to do that.  Humbug, I say.  So I decided to
consolidate that functionality in the generic formatter.

This also had a lot of repeat code, so I made yet another helper to make
testing formatters easier.