Signatures (#106)
* Better error handling with error objects and formatters

The errorResponse module now can make ErrorResponse objects.  If you
pass an ErrorResponse object to `res.send()` it will be formatted as a
standard error output.  To do that, all of the formatters needed to be
rewritten because I could not inject middleware or a prefilter of any
sort into Restify.

This also breaks all of the other formatters out into separate files,
making testers much easier.

Restify also (weirdly) never sets the Content-Length header and relies
on the formatters to do that.  Humbug, I say.  So I decided to
consolidate that functionality in the generic formatter.

This also had a lot of repeat code, so I made yet another helper to make
testing formatters easier.

* Fixing missing period in comment.

* Fixing typo in describe() description.

* Standardizing variable names for formatters.

* Adding signature validation and middleware

This is primarily signature validation.

* Middleware that processes a request and verifies that the basic
  elements for a signed request is present.
* Bulk of work done by version-specific signature module.

Also:

* Changed the mock error response logRef property to say "fakeLogRef" to
  preserve the capitalization of the name.
* New utility functions to pull out the duplicated code.
* OtDate objects are now immutable.  Previously, calling `.plus()` would
  alter the original date object, making it difficult to test for date
  windows or to report on a date after it was modified.
* New HMAC calculation routine in the hash library.

This does *NOT* add the signature validation to any routes.  That is
going to be a separate pull request.  Unfortunately, that also means
that this one isn't functionally tested and there still may be errors
in the software.

* Adding config schema for signatures

* Updating per code review

* Standardizing the variable name

Per code review.

* Better examples

Note: this script has not been tested in practice against the API yet.

* Fixing typo found in code review

* Found a spot during code review that "binary" was ommitted