Guards (#107)
* Fixing format issue

* Not needed with YAML version

* Updating the validation middleware

This renames the middleware to indicate that it only validates the body.
There will be a new middleware that validates the query string, which is
why this change needed to be made.

I also extracted the mock middleware into its own mock file and had to
then update four route tests to inject the mock into the container.

The majority of the rest of the changes are simply updates to use the
new middleware's name.

* Better error handling with error objects and formatters

The errorResponse module now can make ErrorResponse objects.  If you
pass an ErrorResponse object to `res.send()` it will be formatted as a
standard error output.  To do that, all of the formatters needed to be
rewritten because I could not inject middleware or a prefilter of any
sort into Restify.

This also breaks all of the other formatters out into separate files,
making testers much easier.

Restify also (weirdly) never sets the Content-Length header and relies
on the formatters to do that.  Humbug, I say.  So I decided to
consolidate that functionality in the generic formatter.

This also had a lot of repeat code, so I made yet another helper to make
testing formatters easier.

* Making this camel case per code review

* Fixing missing period in comment.

* Fixing typo in describe() description.

* Standardizing variable names for formatters.

* Simple fix to change the output

The variable in the environment is OPENTOKEN_CODE, not OPENTOKEN_KEY and
the parameter is --code instead of --key.

* Adding signature validation and middleware

This is primarily signature validation.

* Middleware that processes a request and verifies that the basic
  elements for a signed request is present.
* Bulk of work done by version-specific signature module.

Also:

* Changed the mock error response logRef property to say "fakeLogRef" to
  preserve the capitalization of the name.
* New utility functions to pull out the duplicated code.
* OtDate objects are now immutable.  Previously, calling `.plus()` would
  alter the original date object, making it difficult to test for date
  windows or to report on a date after it was modified.
* New HMAC calculation routine in the hash library.

This does *NOT* add the signature validation to any routes.  That is
going to be a separate pull request.  Unfortunately, that also means
that this one isn't functionally tested and there still may be errors
in the software.

* Adding config schema for signatures

* Adding guards to managers

This adds functions that ensure that things like an accountId is passed to the managers.  If a required parameter is not passed, this throws an error and that will result in a failed Promise.

Also, for standardization, functions were renamed.

* delAsync -> deleteAsync
* destroyAsync -> deleteAsync

Finally, there are a few bugfixes and adjustments.

* The type checking in chain-middleware.js was fixed to compare against the string "undefined".
* Comments were added to ciphers-and-hashes.js indicating what hashes are known to be insecure.
* If a session is not validated it returns 403 Forbidden instead of 401 Unauthorized.
* The `routesMatch` option for Restify Router Magic needed to be specified because there is a shared bit of code that provides a "GET" response for a couple routes.
* `url` is changed to `uri` because that's closer to right.

* Updating broken tests

* Adding new tests and found a couple missing guards

* Neglected to return the Promise from the test

* Forgot a newline

* Fixing versions of node

* Adjusting test so this works with node 6.5

* Moving guards to storage service

* Adding tests to cover new code

* Missed this one