Various fixes to eliminate errors of multiple types
It was possible to have `.trim()` operate on a non-string before, so I
introduced checks to ensure I have a value instead of `undefined.

Fixed the wording in the documentation.

Updated an example script to have the right permissions and to not read
from stdin any longer.

Fixed where request.body may be a Buffer and it should be changed to a
binary string instead of a UTF-8 encoded string (UTF-8 is the default
when typecasting Buffers to strings).

Ignore Elastic Beanstalk CLI config files.

Added deployment instructions.

Updated config to work as the production config.

Added a specialized health check route.

Removed an unnecessary property from the config.

Prevent nginx from logging requests.

Using Mac-acceptable flags for the date command.