Allowing configuration to be overridden
You can now specify a whole config file by changing the one in the
repository or point at an override file that is merged with the default
settings.  This allows for configurations to be mostly standard and
tweaked to be correct, such as when using the same software for dev, qa,
and production.