Making tests run on Travis, elsewhere
Had to ignore `require()` errors for `bunyan` because it's a dependency
of Restify.  It may or may not be loaded with our `require()` call.  It
should be reloaded with `npm` >= 3.x but does not need to be reloaded
with older versions of `npm` and how it structures the module folder.

Had to reload `node-mocks-http` because it was being required too early.
Now it is always cleared from the cache and reloaded during functional
tests.

`node-mocks-http` uses a `res.header()` function that is slightly
incompatible with Restify.  Submitted issue
howardabrams/node-mocks-http#121 in the hopes that the functionality I
monkey patched will get incorporated into the module directly.

Needed to add `restify-errors` module because we inject it into our
container.  With `npm` < 3.x, it was not available to be loaded via
`require()`.

Remove `util-mock.js` because the real `util.js` does not need
dependencies and it helps test the system by including more working
parts.

Ignore/reset `process.env.DEBUG` and `process.env.PORT` during tests.
Some tests rely on that functionality and I don't want to have them
inconsistently report failures when those environment variables are set.