Need to reset the container
This prevents normal behavior from creeping in during our tests. We want
weird behavior because we're repatching the container's
genericFormatter, but we only want that weird behavior while running one
of these tests.

Before, the genericFormatter was cached automatically by Dizzy and was
reused. I suspect that my recent changes to Dizzy during the enabling of
promises caused the caching technique to work better (in Dizzy's eyes)
and thus break these tests.